Module linkspace::abe

ascii byte expression utilities

ABE is a byte templating language. See guide#ABE to understand its use and indepth explanation. *

Modules

ctx

Custom context for use in varctx

Statics

DEFAULT_PKT

default fmt in many cases and output for [pkt]

Functions

lk_encode

encode bytes as an abe that evaluate back to bytes.

lk_eval

Evaluate an expression and return the bytes

lk_tokenize_abe

An abe parser. Useful to split a cli argument like ’domain:[#:test]:/thing/[12/u32] correctly. The callback is called with (ctrl, contains_brackets, bytes ) where ctrl is 0 | ‘:’ | ‘/’ Only the first ctrl can be ‘\0’.

lk_try_encode

lk_encode with Err on: