Module linkspace::misc

Modules

read

Structs

Cb

RecvPkt

Wrapper around a NetPkt that sets its recv field

ReroutePkt

Wrapper around a netpkt with a mutable NetHeader

ShareArcPkt

Wrapper around a NetPkt that ensures .as_netarc() is only called once.

TreeEntry

A wrapper around TreeKey ( see [[TreeKey::from_fields]] ) and [[TreeValue]] This is the key/value by which packets are saved in the tree index.

Enums

FieldEnum

An enum that provides access to the fields in a NetPkt

StopReason

Statics

DEFAULT_ROUTING_BITS

A thread local default net header value when creating new netpackets

Functions

blake3_hash

Blake3 hash

bytes2uniform

Read bytes as a [0,1) float by reading the first 52 bits. Panics if fewer than 8 bytes are supplied Primary use is to produces the same ‘random’ value by using [NetPkt::hash] or blake3_hash, regardless of language, and without an additional RNG dependencies,

cb

nop_stopped

try_cb