Module linkspace::point

creating points

Functions

lk_datapoint

create a datapoint with upto MAX_CONTENT_SIZE bytes

lk_datapoint_ref

like lk_datapoint but keeps it on the stack in rust enum format.

lk_deserialize

parse a NetPktPtr (the standard binary format) from a buffer

lk_deserialize_len

Determine the length required for lk_deserialize to succeed - expects at least [consts::MIN_NETPKT_LEN] bytes

lk_deserialize_unchecked

like lk_deserialize but skips the hash validation check

lk_keypoint

create a keypoint, i.e. a signed lk_linkpoint

lk_keypoint_ref

like lk_keypoint but keeps it on the stack in rust enum format.

lk_linkpoint

create a new linkpoint NetPktBox

lk_linkpoint_ref

like lk_linkpoint but keeps it on the stack in rust enum format.

lk_serialize

Writes any impl NetPkt into the binary netpkt format