Struct linkspace::prelude::PointTypeFlags

pub struct PointTypeFlags { /* private fields */ }

Pkt flag indicating its type.

Only the _POINT combinations are valid in a packet.

Implementations

impl PointTypeFlags

pub const EMPTY: PointTypeFlags = Self{ bits: 0,}

Indicate that the chances of anybody interested in this packet are zero. Implementations can ignore this, mostly useful for importing many datablocks.

pub const ANY_PKT: PointTypeFlags = Self{ bits: 1,}

pub const DATA: PointTypeFlags = Self{ bits: 1,}

pub const LINK: PointTypeFlags = Self{ bits: 2,}

pub const SIGNATURE: PointTypeFlags = Self{ bits: 4,}

pub const ERROR: PointTypeFlags = Self{ bits: 128,}

pub const DATA_POINT: PointTypeFlags = Self{ bits: Self::DATA.bits,}

pub const LINK_POINT: PointTypeFlags = Self{ bits: Self::DATA.bits | Self::LINK.bits,}

pub const KEY_POINT: PointTypeFlags = Self{ bits: Self::DATA.bits | Self::LINK.bits | Self::SIGNATURE.bits,}

pub const ERROR_POINT: PointTypeFlags = Self{ bits: Self::ERROR.bits,}

pub const fn empty() -> PointTypeFlags

Returns an empty set of flags.

pub const fn all() -> PointTypeFlags

Returns the set containing all flags.

pub const fn bits(&self) -> u8

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u8) -> Option<PointTypeFlags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u8) -> PointTypeFlags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u8) -> PointTypeFlags

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: PointTypeFlags) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: PointTypeFlags) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: PointTypeFlags)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: PointTypeFlags)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: PointTypeFlags)

Toggles the specified flags in-place.

pub fn set(&mut self, other: PointTypeFlags, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: PointTypeFlags) -> PointTypeFlags

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: PointTypeFlags) -> PointTypeFlags

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: PointTypeFlags) -> PointTypeFlags

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the โbit-clearโ operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: PointTypeFlags) -> PointTypeFlags

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> PointTypeFlags

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.

impl PointTypeFlags

pub const fn as_str(self) -> &'static str

pub fn unchecked_from(b: u8) -> PointTypeFlags