Module linkspace::query

Structs

Query

A set of predicates and options used to select packets

Enums

KnownOptions

PredicateType

A list of all supported query predicates

Statics

Q

Functions

lk_hash_query

Create a new Query specifically for a hash. Sets the right mode and ‘i’ count. See [lk_get_hash] if you don’t care to watch the db.

lk_query

Create a new Query. Copy from a template. Q is the empty query.

lk_query_clear

Clear a Query for reuse

lk_query_compile

Compile a Query into a function which tests packets to deteremine if they match - WARN - slow and subject to change.

lk_query_parse

Add multiple ABE encoded statements to a Query

lk_query_print

Get the string representation of a Query

lk_query_push

Add a single statement to a Query, potentially skipping an encode step. i.e. fast path for adding a single statement - lk_query_parse(q,“{field}:{op}:{lk_encode(bytes)}”)