Module linkspace::runtime

Structs

Linkspace

The linkspace runtime.

LkInfo

Functions

lk_get

get the first result from the database matching the query.

lk_get_all

Run callback for every match for the query in the database. Break early if the callback returns true. If break => number of matches If no break ( cb only returned false ) => -1 * Number of matches

lk_get_ref

read a single packet mmap-ed packet. This means that NetPkt::net_header_mut is unavailable. To mutate the header, wrap the result in crate::misc::ReroutePkt or copy with NetPkt::as_netbox.. *

lk_info

lk_list_watches

lk_open

open a linkspace runtime.

lk_process

process the log of new packets and trigger callbacks. Updates the reader to the latest state.

lk_process_while

continuously process callbacks until:

lk_save

save a packet. Returns true if new and false if its old.

lk_save_all

lk_stop

close lk_watch watches based on the query id ‘:qid:example’ in the query.

lk_watch

Before returning, calls cb for every packet in the database. Then registers the query under its ‘qid’ option ( .e.g. set by lk_query_parse(q,“:qid:myqid) ) During a [lk_process] or [lk_process_while] call, for every new packet matching the query, runs the callback.

lk_watch2

lk_watch with a custom log tracing::Span The span will be entered on every callback. If you do not care for structured options you can use vspan lk_watch2(.. , .. ,.. , vspan("my function"))

vspan

See lk_watch2