Module linkspace::misc

misc functions & tools - less stable

Modules

deserialize

Structs

PktCmp

A newtype around any T:NetPkt that implements Eq (and Hash) based on pkt hash and Ord on the standard tree index order.

QuickDedup

Extremely light best-effort deduplication.

RecvPkt

Wrapper around a NetPkt that sets its recv field

ReroutePkt

Wrapper around a netpkt with a mutable NetHeader

ShareArcPkt

Wrapper around a NetPkt that ensures .as_netarc() is only called once.

TreeEntry

‘tree’ index order is a Ord of NetPkts defined by the order of the tuple ( group, domain, space_depth, space, pubkey, create stamp, hash ) A wrapper around TreeKey ( see [[TreeKey::from_fields]] ) and [[TreeValue]] This is the key/value by which packets are saved in the tree index.

Enums

FieldEnum

An enum that provides access to the fields in a NetPkt

Statics

DEFAULT_ROUTING_BITS

A thread local default net header value when creating new netpackets

START_STAMP

Functions

blake3

Blake3 hash

bytes2uniform

Read bytes as a [0,1) float by reading the first 52 bits. Panics if fewer than 8 bytes are supplied Can be used to produces a shared ‘random’ value by using NetPkt::hash or blake3 as input.