Module linkspace::prelude

Re-exports

pub use super::*;

Modules

endian_types

Structs

AB

newtype around bytes to print/parse [[abe]] text

B64

newtype around bytes to print/parse b64 (url-safe no-padding)

Link

A Tag and LkHash

NetFlags

Variable flags used in transit

NetPktArc

Arc around the byte repr NetPkt

NetPktHeader

Variable fields in a crate::NetPkt used in point exchange

NetPktParts

Rust enum repr of a NetPkt

NetPktPtr

Byte repr of a NetPkt

PktFmt

A static packet formatter similar to DEFAULT_PKT without using ABE expressions.

PointTypeFlags

Pkt flag indicating its type.

SigningKey

Enums

PktError

SpaceError

Traits

NetPkt

A trait to access fields of a net pkt. Auto impls NetPktExt, Point, and PointExt.

NetPktExt

Utilities for NetPkt

Point

General trait for accessing point field.

PointExt

Directly access a Point’s domain, group, links, publickey, etc.

SigningExt

Functions

ab

copy val into array of N bytes prepending 0’s as needed. panic if val > N

as_abtxt_c

domain

[Thread Local]: get the ‘default’ domain. from set_domain || $LK_DOMAIN || [0;16]

group

[Thread Local]: get the ‘default’ group. from set_group || $LK_GROUP || [#:pub]

now

current time as big endian u64 microseconds since epoch

rspace1

rspace_buf

set_domain

set the result for domain

set_group

space_buf

try_ab

Type Aliases

Domain

Alias for AB<[u8;16]>

GroupID

Alias for B64<[u8;32]\>

LkHash

Blake3 hash of the packet content. Alias for B64<[u8;32]\>

NetPktBox

Heap allocated repr of a NetPkt.

PubKey

Taproot Schnorr publickey. Alias for B64<[u8;32]>

RecvPktPtr

RootedSpace

RootedSpaceBuf

RootedStaticSpace

Space

Explicitly Space bytes (analogous to [str])

SpaceBuf

Owned Space bytes (analogous to String)

Stamp

A Big endian u64 of microseconds since EPOCH

Tag

Alias for AB<[u8;16]>