Module linkspace::runtime

a runtime to watch for new points from other processes or threads

Modules

cb

(rust only) lk_watch takes the callback PktHandler which are quick to impl with cb and try_cb. Other languages should use their own function syntax as argument to lk_watch. *

Structs

Linkspace

The linkspace runtime.

LkInfo

miscellaneous information about the runtime

Functions

lk_await_write

A future awaiting new lk_write to the storage engine.

lk_info

get LkInfo of a linkspace runtime

lk_last

Last recv stamp of this runtime. i.e. last result from lk_process

lk_list_watches

iterate over all active (Qid,Query)

lk_open

open a linkspace runtime.

lk_open_inmem

opens an in memory linkspace runtime. Unlike lk_open the data is not saved, nor can it be shared across processes.

lk_process

process the log of new packets and trigger callbacks. Updates the reader to the latest state.

lk_process_while

continuously process callbacks until:

lk_read

get the first result from the database matching the query.

lk_read_all

Run callback for every match for the query in the database. Break early if the callback returns true. returns number of matches if return from break return is -1*Number of matches If break => number of matches If no break ( cb only returned false ) => -1 * Number of matches

lk_read_hashes

TODO: DOC

lk_read_ref

read a single packet directly without copying.*

lk_split_runtime

create clone that shares the storage engine, but with an empty runtime,i.e. with its own callbacks and requiring its own lk_process calls.

lk_stop

close lk_watch watches based on the query id ‘:qid:example’ in the query.

lk_threadlocal

get/set the thread local runtime used as default - automatically done for first call to lk_open/lk_open_inmem

lk_watch

Registers the query under its ‘qid’ ( .e.g. set by lk_query_parse(q,“:qid:myqid”) ) Before returning, calls cb for every packet in the database. The absolute return value is the number of times the callback was called. A positive value means the callback finished already.

lk_watch2

lk_watch with a custom log [tracing::Span] The span will be entered on every callback. If you do not care for structured options you can use vspan lk_watch2(.. , .. ,.. , vspan("my function"))

lk_write

save a packet. Returns true if new and false if its old.

lk_write_all

save multiple packets at once - returns the number of new packets written

lk_write_all_ext

returns the range [incusive,exclusive) of recv stamps used to save new packets. total_new = r.1-r.0

vspan

See lk_watch2